Offload some props/state out of ProjectsContainer
