+++ src/app/[[...category]]/ProjectsGrid.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React, { useEffect, useState } from 'react';
 import Grid from '@mui/material/Grid';
